requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
crypto==1.4.1
cryptography==3.4.7
Deprecated==1.2.12
firebase==3.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
gcloud==0.18.3
google-api-core==1.30.0
google-auth==1.32.0
google-cloud==0.34.0
google-cloud-core==1.7.0
google-cloud-storage==1.39.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
greenlet==1.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jwcrypto==0.9.1
MarkupSafe==2.0.1
Naked==0.1.31
oauth2client==4.1.3
openai==0.8.0
packaging==20.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pyparsing==2.4.7
python-jwt==3.3.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.19
sseclient==0.0.27
tqdm==4.61.1
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
